refactor: feature 디렉토리 구조 수정, dependency client Sendable하게 수정
