--- Projects/Data/Sources/SignIn/tmp.swift
@@ -1,6 +1,0 @@
-//
-//  SignInFactoryImplement.swift
-//  Data
-//
-//  Created by 송하민 on 9/16/24.
-//
