+++ Projects/Feature/SignUp/Derived/Sources/TuistBundle+SignUpSampleApp.swift
@@ -4,11 +4,11 @@
 import Foundation// MARK: - Swift Bundle Accessor for Frameworks
 private class BundleFinder {}
 extension Foundation.Bundle {
-/// Since DataSampleApp is a application, the bundle for classes within this module can be used directly.
+/// Since SignUpSampleApp is a application, the bundle for classes within this module can be used directly.
 static let module = Bundle(for: BundleFinder.self)
 }// MARK: - Objective-C Bundle Accessor
 @objc
-public class DataSampleAppResources: NSObject {
+public class SignUpSampleAppResources: NSObject {
 @objc public class var bundle: Bundle {
     return .module
 }
